# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Secret Service
    description: Keptn API for creating secrets
    version: ""
host: localhost:8088
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /openapi.json:
        get:
            tags:
                - openapi
            summary: Download openapi3.json
            operationId: openapi#/openapi.json
            responses:
                "200":
                    description: File downloaded
                    schema:
                        type: file
            schemes:
                - http
    /secret:
        get:
            tags:
                - secret
            summary: getSecret secret
            operationId: secret#getSecret
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/SecretGetSecretResponseBody'
            schemes:
                - http
        post:
            tags:
                - secret
            summary: createSecret secret
            operationId: secret#createSecret
            parameters:
                - name: CreateSecretRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SecretCreateSecretRequestBody'
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/SecretCreateSecretResponseBody'
                "409":
                    description: The secret already exists
                    schema:
                        $ref: '#/definitions/SecretCreateSecretSecretAlreadyExistsResponseBody'
            schemes:
                - http
definitions:
    GetSecretResponseBodyResponseBody:
        title: GetSecretResponseBodyResponseBody
        type: object
        properties:
            keys:
                type: array
                items:
                    type: string
                    example: Sapiente explicabo qui et.
                example:
                    - Eligendi et.
                    - Dolore voluptas.
            name:
                type: string
                example: Quam itaque.
            scope:
                type: string
                example: Ut qui et ab.
        description: Secret result
        example:
            keys:
                - Illo occaecati voluptas velit facilis deserunt molestias.
                - Suscipit qui temporibus similique.
            name: Dolor repellat consequatur molestias.
            scope: Tempore quia sit.
    SecretCreateSecretRequestBody:
        title: SecretCreateSecretRequestBody
        type: object
        properties:
            data:
                type: object
                example:
                    Est maxime eius.: Repellat rerum repellendus.
                    Molestiae et provident quo dignissimos eos.: Aut a iure eaque officiis voluptas praesentium.
                    Non aperiam et.: Dolores consectetur qui.
                additionalProperties:
                    type: string
                    example: Ratione tempore exercitationem dicta.
            name:
                type: string
                example: Repellat iure voluptatem.
            scope:
                type: string
                example: Totam dignissimos ut aut voluptates ab qui.
        example:
            data:
                Eveniet suscipit nobis qui.: Quod quo beatae non.
                Fugit impedit dolorum consequuntur.: Corporis eum maxime architecto qui omnis.
            name: Et molestiae.
            scope: Fugiat et sunt quae veritatis iusto atque.
    SecretCreateSecretResponseBody:
        title: SecretCreateSecretResponseBody
        type: object
        properties:
            data:
                type: object
                example:
                    Aliquam quis qui eum consequatur.: Et ipsam et nobis beatae quis.
                    Dolor eius adipisci.: Sint voluptates ex fugit cupiditate possimus quisquam.
                    Omnis quas quod fuga.: Qui numquam qui.
                additionalProperties:
                    type: string
                    example: Rerum nemo et doloribus aut.
            name:
                type: string
                example: Sed dolores.
            scope:
                type: string
                example: Dolores nobis ipsa autem molestiae iusto.
        example:
            data:
                Dicta quidem.: Ullam magni expedita eum nesciunt.
                Libero aspernatur illo ratione dolore velit.: Modi officia quam architecto dolor laudantium fuga.
                Similique ipsum dolore.: Molestiae nihil doloribus aliquam eveniet.
            name: Sed voluptatem quia molestiae ea omnis.
            scope: Porro nihil velit sit.
    SecretCreateSecretSecretAlreadyExistsResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: true
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: true
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: createSecret_secret_already_exists_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    SecretGetSecretResponseBody:
        title: SecretGetSecretResponseBody
        type: object
        properties:
            secrets:
                type: array
                items:
                    $ref: '#/definitions/GetSecretResponseBodyResponseBody'
                example:
                    - keys:
                        - Perferendis et.
                        - Voluptatem aut.
                        - Illo et autem dolor.
                        - Et suscipit.
                      name: Sit est iusto ipsa.
                      scope: Perferendis numquam voluptate.
                    - keys:
                        - Perferendis et.
                        - Voluptatem aut.
                        - Illo et autem dolor.
                        - Et suscipit.
                      name: Sit est iusto ipsa.
                      scope: Perferendis numquam voluptate.
                    - keys:
                        - Perferendis et.
                        - Voluptatem aut.
                        - Illo et autem dolor.
                        - Et suscipit.
                      name: Sit est iusto ipsa.
                      scope: Perferendis numquam voluptate.
        example:
            secrets:
                - keys:
                    - Perferendis et.
                    - Voluptatem aut.
                    - Illo et autem dolor.
                    - Et suscipit.
                  name: Sit est iusto ipsa.
                  scope: Perferendis numquam voluptate.
                - keys:
                    - Perferendis et.
                    - Voluptatem aut.
                    - Illo et autem dolor.
                    - Et suscipit.
                  name: Sit est iusto ipsa.
                  scope: Perferendis numquam voluptate.
                - keys:
                    - Perferendis et.
                    - Voluptatem aut.
                    - Illo et autem dolor.
                    - Et suscipit.
                  name: Sit est iusto ipsa.
                  scope: Perferendis numquam voluptate.